refactor: move http api from api/ to api/http/ folder
